keepin these for future review
This branch is a bit lost. I want to build something that is actually
portable between systems and C++ is not the best platform to this.
This branch will be resumed to implement the server layer that will
manage the distributed nature of the tasks management.
But that, my friends, is a chapter to be written in the future.